Use a dedicated builder for plugin init containers
